separates row into campaignrow and eventrow
